Enable tslint and resolve errors
