+++ ui/src/flux/components/FilterConditions.tsx
@@ -1,5 +1,6 @@
 import React, {PureComponent} from 'react'
-import {FilterNode, MemberExpressionNode} from 'src/types/flux'
+import {FilterNode} from 'src/types/flux'
+import FilterConditionNode from 'src/flux/components/FilterConditionNode'
 
 interface Props {
   nodes: FilterNode[]
@@ -14,51 +15,6 @@ class FilterConditions extends PureComponent<Props> {
         ))}
       </>
     )
-  }
-}
-
-interface FilterConditionNodeProps {
-  node: FilterNode
-}
-
-/* tslint:disable */
-class FilterConditionNode extends PureComponent<FilterConditionNodeProps> {
-  public render() {
-    const {node} = this.props
-
-    switch (node.type) {
-      case 'ObjectExpression': {
-        return <div className="flux-filter--key">{node.source}</div>
-      }
-      case 'MemberExpression': {
-        const memberNode = node as MemberExpressionNode
-        return (
-          <div className="flux-filter--key">{memberNode.property.name}</div>
-        )
-      }
-      case 'OpenParen': {
-        return <div className="flux-filter--paren-open" />
-      }
-      case 'CloseParen': {
-        return <div className="flux-filter--paren-close" />
-      }
-      case 'NumberLiteral':
-      case 'IntegerLiteral': {
-        return <div className="flux-filter--value number">{node.source}</div>
-      }
-      case 'BooleanLiteral': {
-        return <div className="flux-filter--value boolean">{node.source}</div>
-      }
-      case 'StringLiteral': {
-        return <div className="flux-filter--value string">{node.source}</div>
-      }
-      case 'Operator': {
-        return <div className="flux-filter--operator">{node.source}</div>
-      }
-      default: {
-        return <div />
-      }
-    }
   }
 }
